+++ NEWS.md
@@ -2,8 +2,9 @@
 
 * Define the geographic zones in the documentation
 * Correct the date changes for the values needed to declare pollution emergencies in the README
-* function get_station_imeca for downloading pollution data in IMECAs frome each station
+* function get_station_imeca for downloading pollution data in IMECAs from each station
 * rename get_latest_data to get_latest_imeca and get_zone_data to get_zone_imeca
+* get_station_data now can download TMP, WSP, WDR and RH data back to 1986
 
 # aire.zmvm 0.5.0